+++ +braidlab/@braid/private/sortcross.m
@@ -18,9 +18,8 @@ function crossdat = sortcross(cross_cell)
 %
 %   http://github.com/jeanluct/braidlab
 %
-%   Copyright (C) 2013-2019  Jean-Luc Thiffeault <jeanluc@math.wisc.edu>
+%   Copyright (C) 2013-2021  Jean-Luc Thiffeault <jeanluc@math.wisc.edu>
 %                            Marko Budisic          <marko@clarkson.edu>
-%                       Michael Allshouse <m.allshouse@northeastern.edu>
 %
 %   This file is part of Braidlab.
 %