+++ +braidlab/@braid/private/sortcross2gen.m
@@ -15,9 +15,8 @@ function [gen,tcr] = sortcross2gen(n,crossdat)
 %
 %   http://github.com/jeanluct/braidlab
 %
-%   Copyright (C) 2013-2019  Jean-Luc Thiffeault <jeanluc@math.wisc.edu>
+%   Copyright (C) 2013-2021  Jean-Luc Thiffeault <jeanluc@math.wisc.edu>
 %                            Marko Budisic          <marko@clarkson.edu>
-%                       Michael Allshouse <m.allshouse@northeastern.edu>
 %
 %   This file is part of Braidlab.
 %